backend/requirements.txt
# Backend Dependencies for AI Chatbot
# FastAPI framework and server
fastapi==0.104.1
uvicorn==0.24.0

# Environment variables management
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# CORS support
python-multipart==0.0.6

# AI Model Integration (choose one or more based on your preference)
# Option 1: OpenAI (uncomment if using OpenAI GPT)
openai==1.3.0

# Option 2: Hugging Face Transformers (uncomment if using local models)
# transformers==4.35.0
# torch==2.1.0

# Option 3: Ollama (uses HTTP requests, no special package needed)
# requests==2.31.0

# General HTTP requests (useful for any API integration)
requests==2.31.0
httpx==0.25.2

youtube-transcript-api>=0.6.3
pytube==15.0.0